+++ node_backend/src/models/TenantDeviceType.ts
@@ -7,6 +7,7 @@ export interface TenantDeviceType {
   tenantDeviceCategoryId: number;
   name: string;
   code: string;
+  widgetUuid: string | null;
   isModelNumberRequired: boolean;
   isSerialNumberRequired: boolean;
   isMacAddressRequired: boolean;
@@ -24,6 +25,7 @@ export interface TenantDeviceType {
 export interface CreateDeviceTypeDTO {
   tenantDeviceCategoryId: number;
   name: string;
+  widgetUuid?: string | null;
   isModelNumberRequired?: boolean;
   isSerialNumberRequired?: boolean;
   isMacAddressRequired?: boolean;
@@ -35,6 +37,7 @@ export interface CreateDeviceTypeDTO {
 export interface UpdateDeviceTypeDTO {
   tenantDeviceCategoryId?: number;
   name?: string;
+  widgetUuid?: string | null;
   isModelNumberRequired?: boolean;
   isSerialNumberRequired?: boolean;
   isMacAddressRequired?: boolean;